+++ CHANGELOG.md
@@ -16,12 +16,14 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ### Added
 
-* Weight refresh mechanism for ``OneSidedUnitCell`` to counteract
-  saturation, by differential read, reset, and re-write. (\#209)
 * A number of new config presets added to the library, namely `EcRamMOPreset`,
   `EcRamMO2Preset`, `EcRamMO4Preset`, `TikiTakaEcRamMOPreset`,
   `MixedPrecisionEcRamMOPreset`. These can be used for tile configuration
   (`rpu_config`). They specify a particular device and optimizer choice. (\#207)
+* Weight refresh mechanism for ``OneSidedUnitCell`` to counteract
+  saturation, by differential read, reset, and re-write. (\#209)
+* Complex cycle-to-cycle noise for ``ExpStepDevice`` (\#???)
+* PCM Preset added (\#???)
 
 ### Changed